% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemblynew.R
\docType{data}
\name{assemblynew}
\alias{assemblynew}
\title{assemblynew}
\format{
An object of class \code{data.frame} with 16 rows and 5 columns.
}
\usage{
assemblynew
}
\description{
Problem 4.38
}
\keyword{datasets}
