% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballbearings.R
\docType{data}
\name{ballbearings}
\alias{ballbearings}
\title{ballbearings}
\format{
An object of class \code{data.frame} with 24 rows and 2 columns.
}
\usage{
ballbearings
}
\description{
Problem 2.29
}
\keyword{datasets}
