% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batteries.R
\docType{data}
\name{batteries}
\alias{batteries}
\title{batteries}
\format{
An object of class \code{data.frame} with 15 rows and 2 columns.
}
\usage{
batteries
}
\description{
Problem 3.27S
}
\keyword{datasets}
