% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batteriesdevices.R
\docType{data}
\name{batteriesdevices}
\alias{batteriesdevices}
\title{batteriesdevices}
\format{
An object of class \code{data.frame} with 12 rows and 3 columns.
}
\usage{
batteriesdevices
}
\description{
Problem 5.27
}
\keyword{datasets}
