% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity.R
\docType{data}
\name{bioactivity}
\alias{bioactivity}
\title{bioactivity}
\format{
An object of class \code{data.frame} with 12 rows and 3 columns.
}
\usage{
bioactivity
}
\description{
Problem 3.14S
}
\keyword{datasets}
