% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boltstrength.R
\docType{data}
\name{boltstrength}
\alias{boltstrength}
\title{boltstrength}
\format{
An object of class \code{data.frame} with 20 rows and 3 columns.
}
\usage{
boltstrength
}
\description{
Problem 4.7S
}
\keyword{datasets}
