% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boneanchors.R
\docType{data}
\name{boneanchors}
\alias{boneanchors}
\title{boneanchors}
\format{
An object of class \code{data.frame} with 12 rows and 3 columns.
}
\usage{
boneanchors
}
\description{
Problem 5.39
}
\keyword{datasets}
