% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellphones.R
\docType{data}
\name{cellphones}
\alias{cellphones}
\title{cellphones}
\format{
An object of class \code{data.frame} with 40 rows and 2 columns.
}
\usage{
cellphones
}
\description{
Problem 2.28S
}
\keyword{datasets}
