% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical6.R
\docType{data}
\name{chemical6}
\alias{chemical6}
\title{chemical6}
\format{
An object of class \code{data.frame} with 32 rows and 5 columns.
}
\usage{
chemical6
}
\description{
Problem 6.11
}
\keyword{datasets}
