% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuitnoise.R
\docType{data}
\name{circuitnoise}
\alias{circuitnoise}
\title{circuitnoise}
\format{
An object of class \code{data.frame} with 20 rows and 2 columns.
}
\usage{
circuitnoise
}
\description{
Problem 3.25
}
\keyword{datasets}
