% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colortubes.R
\docType{data}
\name{colortubes}
\alias{colortubes}
\title{colortubes}
\format{
An object of class \code{data.frame} with 16 rows and 2 columns.
}
\usage{
colortubes
}
\description{
Problem 3.20
}
\keyword{datasets}
