% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/developmentcost.R
\docType{data}
\name{developmentcost}
\alias{developmentcost}
\title{developmentcost}
\format{
An object of class \code{data.frame} with 36 rows and 3 columns.
}
\usage{
developmentcost
}
\description{
Problem 4.14
}
\keyword{datasets}
