% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddversion.R
\name{eddversion}
\alias{eddversion}
\title{eddversion}
\usage{
eddversion()
}
\value{
The version number as a package version object.
}
\description{
Get the version number of the expdesigndata (Experimental Design Data) package. eddversion takes no arguments.
}
