% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillheight.R
\docType{data}
\name{fillheight}
\alias{fillheight}
\title{fillheight}
\format{
An object of class \code{data.frame} with 16 rows and 4 columns.
}
\usage{
fillheight
}
\description{
Problem 6.20
}
\keyword{datasets}
