% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardwoodpaper.R
\docType{data}
\name{hardwoodpaper}
\alias{hardwoodpaper}
\title{hardwoodpaper}
\format{
An object of class \code{data.frame} with 36 rows and 4 columns.
}
\usage{
hardwoodpaper
}
\description{
Problem 5.37
}
\keyword{datasets}
