% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insulatingfluid.R
\docType{data}
\name{insulatingfluid}
\alias{insulatingfluid}
\title{insulatingfluid}
\format{
An object of class \code{data.frame} with 24 rows and 2 columns.
}
\usage{
insulatingfluid
}
\description{
Problem 3.24S
}
\keyword{datasets}
