% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyboards.R
\docType{data}
\name{keyboards}
\alias{keyboards}
\title{keyboards}
\format{
An object of class \code{data.frame} with 18 rows and 3 columns.
}
\usage{
keyboards
}
\description{
Problem 5.40
}
\keyword{datasets}
