% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mailmarketing.R
\docType{data}
\name{mailmarketing}
\alias{mailmarketing}
\title{mailmarketing}
\format{
An object of class \code{data.frame} with 12 rows and 3 columns.
}
\usage{
mailmarketing
}
\description{
Problem 4.12S
}
\keyword{datasets}
