% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newmaterial.R
\docType{data}
\name{newmaterial}
\alias{newmaterial}
\title{newmaterial}
\format{
An object of class \code{data.frame} with 25 rows and 4 columns.
}
\usage{
newmaterial
}
\description{
Problem 4.24
}
\keyword{datasets}
