% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operadonations.R
\docType{data}
\name{operadonations}
\alias{operadonations}
\title{operadonations}
\format{
An object of class \code{data.frame} with 24 rows and 3 columns.
}
\usage{
operadonations
}
\description{
Problem 3.17
}
\keyword{datasets}
