% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscilloscope.R
\docType{data}
\name{oscilloscope}
\alias{oscilloscope}
\title{oscilloscope}
\format{
An object of class \code{data.frame} with 27 rows and 3 columns.
}
\usage{
oscilloscope
}
\description{
Problem 5.12
}
\keyword{datasets}
