% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphorblock.R
\docType{data}
\name{phosphorblock}
\alias{phosphorblock}
\title{phosphorblock}
\format{
An object of class \code{data.frame} with 18 rows and 4 columns.
}
\usage{
phosphorblock
}
\description{
Problem 5.35
}
\keyword{datasets}
