% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasticpipe.R
\docType{data}
\name{plasticpipe}
\alias{plasticpipe}
\title{plasticpipe}
\format{
An object of class \code{data.frame} with 24 rows and 2 columns.
}
\usage{
plasticpipe
}
\description{
Problem 2.32
}
\keyword{datasets}
