% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polysilicon.R
\docType{data}
\name{polysilicon}
\alias{polysilicon}
\title{polysilicon}
\format{
An object of class \code{data.frame} with 12 rows and 3 columns.
}
\usage{
polysilicon
}
\description{
Problem 5.26
}
\keyword{datasets}
