% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processdev.R
\docType{data}
\name{processdev}
\alias{processdev}
\title{processdev}
\format{
An object of class \code{data.frame} with 16 rows and 5 columns.
}
\usage{
processdev
}
\description{
Problem 6.26
}
\keyword{datasets}
