% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiocontrol.R
\docType{data}
\name{radiocontrol}
\alias{radiocontrol}
\title{radiocontrol}
\format{
An object of class \code{data.frame} with 24 rows and 4 columns.
}
\usage{
radiocontrol
}
\description{
Problem 4.16
}
\keyword{datasets}
