% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siliconetching.R
\docType{data}
\name{siliconetching}
\alias{siliconetching}
\title{siliconetching}
\format{
An object of class \code{data.frame} with 16 rows and 2 columns.
}
\usage{
siliconetching
}
\description{
Problem 2.34S
}
\keyword{datasets}
