% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silverpowder.R
\docType{data}
\name{silverpowder}
\alias{silverpowder}
\title{silverpowder}
\format{
An object of class \code{data.frame} with 16 rows and 5 columns.
}
\usage{
silverpowder
}
\description{
Problem 5.31
}
\keyword{datasets}
