% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thickness.R
\docType{data}
\name{thickness}
\alias{thickness}
\title{thickness}
\format{
An object of class \code{data.frame} with 16 rows and 2 columns.
}
\usage{
thickness
}
\description{
Problem 2.27
}
\keyword{datasets}
