% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toc.R
\name{toc}
\alias{toc}
\title{toc}
\usage{
toc()
}
\description{
\code{help(toc)} displays the table of contents for datasets.
}
\details{
A guide to the datasets in expdesign. Datasets are listed under the Arguments section by the order they appear in the Nonparametric Statistical Methods textbook, 3ed. The reference name in the \code{nsm3data} package is given, along with the table number and the descriptor of the dataset.

\tabular{ll}{
\bold{Table} \tab \bold{Table Number - Description} \cr
\code{shelflife} \tab {Problem 2.19S} \cr
\code{hours} \tab {Problem 2.21} \cr
\code{machines} \tab {Problem 2.23} \cr
\code{flares} \tab {Problem 2.25} \cr
\code{plasma} \tab {Problem 2.25} \cr
\code{thickness} \tab {Problem 2.27} \cr
\code{cellphones} \tab {Problem 2.28S} \cr
}
}
\examples{
\dontrun{
help(toc)
}

}
