% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twofactexp.R
\docType{data}
\name{twofactexp}
\alias{twofactexp}
\title{twofactexp}
\format{
An object of class \code{data.frame} with 12 rows and 3 columns.
}
\usage{
twofactexp
}
\description{
Problem 5.17
}
\keyword{datasets}
