% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valveshutoff.R
\docType{data}
\name{valveshutoff}
\alias{valveshutoff}
\title{valveshutoff}
\format{
An object of class \code{data.frame} with 15 rows and 2 columns.
}
\usage{
valveshutoff
}
\description{
Problem 3.23
}
\keyword{datasets}
