% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virusgrowth.R
\docType{data}
\name{virusgrowth}
\alias{virusgrowth}
\title{virusgrowth}
\format{
An object of class \code{data.frame} with 24 rows and 3 columns.
}
\usage{
virusgrowth
}
\description{
Problem 6.12
}
\keyword{datasets}
