% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waferparticles.R
\docType{data}
\name{waferparticles}
\alias{waferparticles}
\title{waferparticles}
\format{
An object of class \code{data.frame} with 15 rows and 2 columns.
}
\usage{
waferparticles
}
\description{
Problem 3.29
}
\keyword{datasets}
