% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waferthickness.R
\docType{data}
\name{waferthickness}
\alias{waferthickness}
\title{waferthickness}
\format{
An object of class \code{data.frame} with 12 rows and 2 columns.
}
\usage{
waferthickness
}
\description{
Problem 3.31
}
\keyword{datasets}
