% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toc.R
\name{toc}
\alias{toc}
\title{toc}
\usage{
toc()
}
\description{
\code{help(toc)} displays the table of contents for datasets.
}
\details{
A guide to the datasets in expdesign. Datasets are listed under the Arguments section by the order they appear in the Experimental Design textbook, 10ed. The reference name in the \code{expdesigndata} package is given, along with the table number and the descriptor of the dataset.

\tabular{ll}{
\bold{Table} \tab \bold{Table Number - Description} \cr
\code{shelflife} \tab {Problem 2.19S} \cr
\code{hours} \tab {Problem 2.21} \cr
\code{machines} \tab {Problem 2.23} \cr
\code{flares} \tab {Problem 2.25} \cr
\code{plasma} \tab {Problem 2.26S} \cr
\code{thickness} \tab {Problem 2.27} \cr
\code{cellphones} \tab {Problem 2.28S} \cr
\code{ballbearings} \tab {Problem 2.29} \cr
\code{births} \tab {Problem 2.30} \cr
\code{girders} \tab {Problem 2.31S} \cr
\code{plasticpipe} \tab {Problem 2.32} \cr
\code{siliconetching} \tab {Problem 2.34S} \cr
\code{concrete} \tab {Problem 3.19} \cr
\code{cotton} \tab {Problem 3.12} \cr
\code{bioactivity} \tab {Problem 3.14S} \cr
\code{cars} \tab {Problem 3.15} \cr
\code{golf} \tab {Problem 3.16} \cr
\code{operadonations} \tab {Problem 3.17} \cr
\code{bricks} \tab {Problem 3.18} \cr
\code{colortubes} \tab {Problem 3.20} \cr
\code{radon} \tab {Problem 3.22S} \cr
\code{valveshutoff} \tab {Problem 3.23} \cr
\code{insulatingfluid} \tab {Problem 3.24S} \cr
\code{circuitnoise} \tab {Problem 3.25} \cr
\code{methyl} \tab {Problem 3.26} \cr
\code{batteries} \tab {Problem 3.27S} \cr
\code{catalyst} \tab {Problem 3.28} \cr
\code{waferparticles} \tab {Problem 3.29} \cr
\code{ovens} \tab {Problem 3.30S} \cr
\code{waferthickness} \tab {Problem 3.31} \cr
\code{cncfeed} \tab {Problem 3.49} \cr
\code{boltstrength} \tab {Problem 4.7S} \cr
\code{milk} \tab {Problem 4.8} \cr
\code{coupon} \tab {Problem 4.11} \cr
\code{mailmarketing} \tab {Problem 4.12S} \cr
\code{truck} \tab {Problem 4.13} \cr
\code{developmentcost} \tab {Problem 4.14} \cr
\code{leukemia} \tab {Problem 4.15} \cr
\code{radiocontrol} \tab {Problem 4.16} \cr
\code{aluminum} \tab {Problem 4.17S} \cr
\code{eyefocus} \tab {Problem 4.23} \cr
\code{newmaterial} \tab {Problem 4.24} \cr
\code{assembly} \tab {Problem 4.25S} \cr
\code{acid} \tab {Problem 4.37} \cr
\code{assemblynew} \tab {Problem 4.38} \cr
\code{additive} \tab {Problem 4.42S} \cr
\code{hardwood} \tab {Problem 4.44S} \cr
\code{chemical} \tab {Problem 5.7} \cr
\code{surface} \tab {Problem 5.8} \cr
\code{phosphor} \tab {Problem 5.10S} \cr
\code{fiber} \tab {Problem 5.11S} \cr
\code{oscilloscope} \tab {Problem 5.12} \cr
\code{carbon} \tab {Problem 5.15S} \cr
\code{twofactexp} \tab {Problem 5.17} \cr
\code{adhesive} \tab {Problem 5.18} \cr
\code{hardwoodpaper} \tab {Problem 5.20} \cr
\code{fabric} \tab {Problem 5.21} \cr
\code{fatigue} \tab {Problem 5.25S} \cr
\code{polysilicon} \tab {Problem 5.26} \cr
\code{batteriesdevices} \tab {Problem 5.27} \cr
\code{stains} \tab {Problem 5.28} \cr
\code{boneanchors} \tab {Problem 5.29} \cr
\code{keyboards} \tab {Problem 5.30} \cr
\code{silverpowder} \tab {Problem 5.31} \cr
\code{nisin} \tab {Problem 5.33} \cr
\code{phosphorblock} \tab {Problem 5.35} \cr
\code{fiber} \tab {Problem 5.36} \cr
\code{hardwoodpaper} \tab {Problem 5.37} \cr
\code{fabric} \tab {Problem 5.38S} \cr
\code{boneanchors} \tab {Problem 5.39} \cr
\code{keyboards} \tab {Problem 5.40} \cr
\code{chemical6} \tab {Problem 6.11} \cr
\code{virusgrowth} \tab {Problem 6.12} \cr
\code{processdev} \tab {Problem 6.26} \cr
}
}
\examples{
\dontrun{
help(toc)
}

}
